50/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
yarl==1.7.2
